remove LALInspiralWaveForInjection file and add the function in LALInspiralWave. Get rid of warnings as well Original: f5fafca7ece51cd75454533c806638c94944b7f9

+++ lal/packages/inspiral/src/LALInspiralWave1.c
@@ -595,7 +595,7 @@ LALInspiralWave1ForInjection(
 
   if ( (p/2./LAL_PI) < 2 ){
     sprintf(message, "The waveform has only %lf cycles; we don't keep waveform with less than 2 cycles.", 
-	       p/2./LAL_PI );
+	       p/2./(double)LAL_PI );
     LALWarning(status, message);
   }
   else {
